Updated to dotnet core 7 and to MS SQL Server to be able to update to EF core 7

+++ MundoBalloonApi.business/DTOs/Entities/EventCartDetail.cs
@@ -1,13 +1,13 @@
 ﻿namespace MundoBalloonApi.business.DTOs.Entities;
 
-public class EventCartDetail: BaseDto
+public class EventCartDetail : BaseDto
 {
     public int EventCartId { get; set; }
     public string Sku { get; set; } = string.Empty;
     public int ProductVariantId { get; set; }
     public int? UserEventId { get; init; }
-    public double Quantity { get; set; }
-    public double Price { get; set; }
+    public decimal Quantity { get; set; }
+    public decimal Price { get; set; }
     public string Label { get; set; } = string.Empty;
 
     public ProductVariant? Variant { get; set; }
