+++ MundoBalloonApi.business/DTOs/Entities/ProductVariantReview.cs
@@ -5,9 +5,10 @@ public class ProductVariantReview : BaseDto
     public int ProductVariantReviewId { get; init; }
     public int ProductVariantId { get; init; }
     public string UserId { get; init; } = string.Empty;
-    public int Rating { get; init; }
+    public short Rating { get; init; }
     public string? Comments { get; init; }
 
     public ProductVariant? ProductVariant { get; set; }
+
     public User? User { get; set; }
 }
